src/backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
diff-match-patch==20230430
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-import-export==4.1.1
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.0.8
openpyxl==3.1.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
redis==5.0.8
service-identity==24.1.0
sqlparse==0.5.1
tablib==3.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
zope.interface==7.0.3
